Move execution requests processing to ExecutionRequestsProcessor
