+++ ethereum/spec/src/main/java/tech/pegasys/teku/spec/logic/common/block/BlockProcessor.java
@@ -14,7 +14,6 @@
 package tech.pegasys.teku.spec.logic.common.block;
 
 import it.unimi.dsi.fastutil.objects.Object2IntMap;
-import java.util.List;
 import java.util.Optional;
 import java.util.function.Supplier;
 import tech.pegasys.teku.bls.BLSPublicKey;
@@ -32,9 +31,6 @@ import tech.pegasys.teku.spec.datastructures.execution.ExecutionPayloadHeader;
 import tech.pegasys.teku.spec.datastructures.execution.ExecutionPayloadSummary;
 import tech.pegasys.teku.spec.datastructures.execution.ExpectedWithdrawals;
 import tech.pegasys.teku.spec.datastructures.execution.NewPayloadRequest;
-import tech.pegasys.teku.spec.datastructures.execution.versions.electra.ConsolidationRequest;
-import tech.pegasys.teku.spec.datastructures.execution.versions.electra.DepositRequest;
-import tech.pegasys.teku.spec.datastructures.execution.versions.electra.WithdrawalRequest;
 import tech.pegasys.teku.spec.datastructures.operations.Attestation;
 import tech.pegasys.teku.spec.datastructures.operations.AttestationData;
 import tech.pegasys.teku.spec.datastructures.operations.AttesterSlashing;
@@ -167,23 +163,6 @@ public interface BlockProcessor {
       throws BlockProcessingException;
 
   void processWithdrawals(MutableBeaconState state, ExecutionPayloadSummary payloadSummary)
-      throws BlockProcessingException;
-
-  void processDepositRequests(MutableBeaconState state, List<DepositRequest> depositRequests)
-      throws BlockProcessingException;
-
-  void processWithdrawalRequests(
-      MutableBeaconState state,
-      List<WithdrawalRequest> exits,
-      Supplier<ValidatorExitContext> validatorExitContextSupplier)
-      throws BlockProcessingException;
-
-  void processConsolidationRequests(
-      MutableBeaconState state, List<ConsolidationRequest> consolidationRequests)
-      throws BlockProcessingException;
-
-  boolean isValidSwitchToCompoundingRequest(
-      BeaconState beaconState, ConsolidationRequest consolidationRequest)
       throws BlockProcessingException;
 
   ExpectedWithdrawals getExpectedWithdrawals(BeaconState preState);
