+++ ethereum/spec/src/main/java/tech/pegasys/teku/spec/logic/versions/electra/block/BlockProcessorElectra.java
@@ -15,24 +15,18 @@ package tech.pegasys.teku.spec.logic.versions.electra.block;
 
 import static com.google.common.base.Preconditions.checkArgument;
 import static tech.pegasys.teku.infrastructure.unsigned.UInt64.ZERO;
-import static tech.pegasys.teku.spec.config.SpecConfig.FAR_FUTURE_EPOCH;
-import static tech.pegasys.teku.spec.config.SpecConfigElectra.FULL_EXIT_REQUEST_AMOUNT;
 
 import it.unimi.dsi.fastutil.ints.IntList;
 import it.unimi.dsi.fastutil.objects.Object2IntMap;
 import java.util.List;
+import java.util.NoSuchElementException;
 import java.util.Optional;
 import java.util.function.Supplier;
 import java.util.stream.IntStream;
-import org.apache.logging.log4j.LogManager;
-import org.apache.logging.log4j.Logger;
 import org.apache.tuweni.bytes.Bytes32;
 import tech.pegasys.teku.bls.BLSPublicKey;
 import tech.pegasys.teku.bls.BLSSignature;
-import tech.pegasys.teku.ethereum.execution.types.Eth1Address;
-import tech.pegasys.teku.infrastructure.bytes.Bytes20;
 import tech.pegasys.teku.infrastructure.ssz.SszList;
-import tech.pegasys.teku.infrastructure.ssz.SszMutableList;
 import tech.pegasys.teku.infrastructure.ssz.collections.SszBitlist;
 import tech.pegasys.teku.infrastructure.ssz.primitive.SszBytes32;
 import tech.pegasys.teku.infrastructure.ssz.primitive.SszUInt64;
@@ -45,19 +39,14 @@ import tech.pegasys.teku.spec.datastructures.execution.ExecutionPayload;
 import tech.pegasys.teku.spec.datastructures.execution.ExecutionPayloadSummary;
 import tech.pegasys.teku.spec.datastructures.execution.ExpectedWithdrawals;
 import tech.pegasys.teku.spec.datastructures.execution.NewPayloadRequest;
-import tech.pegasys.teku.spec.datastructures.execution.versions.electra.ConsolidationRequest;
-import tech.pegasys.teku.spec.datastructures.execution.versions.electra.DepositRequest;
 import tech.pegasys.teku.spec.datastructures.execution.versions.electra.ExecutionRequests;
 import tech.pegasys.teku.spec.datastructures.execution.versions.electra.ExecutionRequestsDataCodec;
-import tech.pegasys.teku.spec.datastructures.execution.versions.electra.WithdrawalRequest;
 import tech.pegasys.teku.spec.datastructures.operations.Attestation;
 import tech.pegasys.teku.spec.datastructures.operations.Deposit;
-import tech.pegasys.teku.spec.datastructures.state.Validator;
 import tech.pegasys.teku.spec.datastructures.state.beaconstate.BeaconState;
 import tech.pegasys.teku.spec.datastructures.state.beaconstate.MutableBeaconState;
 import tech.pegasys.teku.spec.datastructures.state.beaconstate.versions.electra.BeaconStateElectra;
 import tech.pegasys.teku.spec.datastructures.state.beaconstate.versions.electra.MutableBeaconStateElectra;
-import tech.pegasys.teku.spec.datastructures.state.versions.electra.PendingConsolidation;
 import tech.pegasys.teku.spec.datastructures.state.versions.electra.PendingDeposit;
 import tech.pegasys.teku.spec.datastructures.type.SszKZGCommitment;
 import tech.pegasys.teku.spec.datastructures.type.SszPublicKey;
@@ -75,22 +64,19 @@ import tech.pegasys.teku.spec.logic.common.util.SyncCommitteeUtil;
 import tech.pegasys.teku.spec.logic.common.util.ValidatorsUtil;
 import tech.pegasys.teku.spec.logic.versions.deneb.block.BlockProcessorDeneb;
 import tech.pegasys.teku.spec.logic.versions.deneb.types.VersionedHash;
+import tech.pegasys.teku.spec.logic.versions.electra.execution.ExecutionRequestsProcessorElectra;
 import tech.pegasys.teku.spec.logic.versions.electra.helpers.BeaconStateAccessorsElectra;
 import tech.pegasys.teku.spec.logic.versions.electra.helpers.BeaconStateMutatorsElectra;
 import tech.pegasys.teku.spec.logic.versions.electra.helpers.MiscHelpersElectra;
-import tech.pegasys.teku.spec.logic.versions.electra.helpers.PredicatesElectra;
 import tech.pegasys.teku.spec.schemas.SchemaDefinitionsElectra;
 
 public class BlockProcessorElectra extends BlockProcessorDeneb {
 
-  private static final Logger LOG = LogManager.getLogger();
-
   private final SpecConfigElectra specConfigElectra;
-  private final PredicatesElectra predicatesElectra;
-  private final BeaconStateMutatorsElectra beaconStateMutatorsElectra;
   private final BeaconStateAccessorsElectra beaconStateAccessorsElectra;
   private final SchemaDefinitionsElectra schemaDefinitionsElectra;
   private final ExecutionRequestsDataCodec executionRequestsDataCodec;
+  private final ExecutionRequestsProcessorElectra executionRequestsProcessorElectra;
 
   public BlockProcessorElectra(
       final SpecConfigElectra specConfig,
@@ -105,7 +91,8 @@ public class BlockProcessorElectra extends BlockProcessorDeneb {
       final ValidatorsUtil validatorsUtil,
       final OperationValidator operationValidator,
       final SchemaDefinitionsElectra schemaDefinitions,
-      final ExecutionRequestsDataCodec executionRequestsDataCodec) {
+      final ExecutionRequestsDataCodec executionRequestsDataCodec,
+      final ExecutionRequestsProcessorElectra executionRequestsProcessorElectra) {
     super(
         specConfig,
         predicates,
@@ -120,11 +107,10 @@ public class BlockProcessorElectra extends BlockProcessorDeneb {
         operationValidator,
         schemaDefinitions);
     this.specConfigElectra = specConfig;
-    this.predicatesElectra = PredicatesElectra.required(predicates);
-    this.beaconStateMutatorsElectra = beaconStateMutators;
     this.beaconStateAccessorsElectra = beaconStateAccessors;
     this.schemaDefinitionsElectra = schemaDefinitions;
     this.executionRequestsDataCodec = executionRequestsDataCodec;
+    this.executionRequestsProcessorElectra = executionRequestsProcessorElectra;
   }
 
   @Override
@@ -160,17 +146,24 @@ public class BlockProcessorElectra extends BlockProcessorDeneb {
     super.processOperationsNoValidation(
         state, body, indexedAttestationCache, validatorExitContextSupplier);
 
-    safelyProcess(
-        () -> {
-          final ExecutionRequests executionRequests =
-              body.getOptionalExecutionRequests()
-                  .orElseThrow(() -> new BlockProcessingException("Execution requests expected"));
+    safelyProcess(() -> processExecutionRequests(state, body, validatorExitContextSupplier));
+  }
 
-          processDepositRequests(state, executionRequests.getDeposits());
-          processWithdrawalRequests(
-              state, executionRequests.getWithdrawals(), validatorExitContextSupplier);
-          processConsolidationRequests(state, executionRequests.getConsolidations());
-        });
+  public void processExecutionRequests(
+      final MutableBeaconState state,
+      final BeaconBlockBody body,
+      final Supplier<ValidatorExitContext> validatorExitContextSupplier) {
+    final ExecutionRequests executionRequests =
+        body.getOptionalExecutionRequests()
+            .orElseThrow(
+                () -> new NoSuchElementException("Execution requests expected as part of body"));
+
+    executionRequestsProcessorElectra.processDepositRequests(
+        state, executionRequests.getDeposits());
+    executionRequestsProcessorElectra.processWithdrawalRequests(
+        state, executionRequests.getWithdrawals(), validatorExitContextSupplier);
+    executionRequestsProcessorElectra.processConsolidationRequests(
+        state, executionRequests.getConsolidations());
   }
 
   @Override
@@ -210,408 +203,6 @@ public class BlockProcessorElectra extends BlockProcessorDeneb {
         schemaDefinitionsElectra,
         beaconStateMutators,
         specConfigElectra);
-  }
-
-  /*
-   Implements process_deposit_request from consensus-specs (EIP-6110)
-  */
-  @Override
-  public void processDepositRequests(
-      final MutableBeaconState state, final List<DepositRequest> depositRequests) {
-    final MutableBeaconStateElectra electraState = MutableBeaconStateElectra.required(state);
-    final SszMutableList<PendingDeposit> pendingDeposits =
-        MutableBeaconStateElectra.required(state).getPendingDeposits();
-    for (DepositRequest depositRequest : depositRequests) {
-      // process_deposit_request
-      if (electraState
-          .getDepositRequestsStartIndex()
-          .equals(SpecConfigElectra.UNSET_DEPOSIT_REQUESTS_START_INDEX)) {
-        electraState.setDepositRequestsStartIndex(depositRequest.getIndex());
-      }
-
-      final PendingDeposit deposit =
-          schemaDefinitionsElectra
-              .getPendingDepositSchema()
-              .create(
-                  new SszPublicKey(depositRequest.getPubkey()),
-                  SszBytes32.of(depositRequest.getWithdrawalCredentials()),
-                  SszUInt64.of(depositRequest.getAmount()),
-                  new SszSignature(depositRequest.getSignature()),
-                  SszUInt64.of(state.getSlot()));
-      pendingDeposits.append(deposit);
-    }
-  }
-
-  /** Implements process_withdrawal_request from consensus-specs (EIP-7002 & EIP-7251). */
-  @Override
-  public void processWithdrawalRequests(
-      final MutableBeaconState state,
-      final List<WithdrawalRequest> withdrawalRequests,
-      final Supplier<ValidatorExitContext> validatorExitContextSupplier) {
-    final UInt64 slot = state.getSlot();
-    final UInt64 currentEpoch = miscHelpers.computeEpochAtSlot(slot);
-
-    LOG.debug(
-        "process_withdrawal_request: {} withdrawal request to process from block at " + "slot {}",
-        withdrawalRequests.size(),
-        slot);
-
-    withdrawalRequests.forEach(
-        withdrawalRequest -> {
-          LOG.debug(
-              "process_withdrawal_request: processing withdrawal request {}", withdrawalRequest);
-
-          // If partial withdrawal queue is full, only full exits are processed
-          final boolean isFullExitRequest =
-              withdrawalRequest.getAmount().equals(FULL_EXIT_REQUEST_AMOUNT);
-          final boolean partialWithdrawalsQueueFull =
-              state.toVersionElectra().orElseThrow().getPendingPartialWithdrawals().size()
-                  == specConfigElectra.getPendingPartialWithdrawalsLimit();
-          if (partialWithdrawalsQueueFull && !isFullExitRequest) {
-            LOG.debug("process_withdrawal_request: partial withdrawal queue is full");
-            return;
-          }
-
-          final Optional<Integer> maybeValidatorIndex =
-              validatorsUtil.getValidatorIndex(state, withdrawalRequest.getValidatorPubkey());
-          if (maybeValidatorIndex.isEmpty()) {
-            LOG.debug(
-                "process_withdrawal_request: no matching validator for public key {}",
-                withdrawalRequest.getValidatorPubkey().toAbbreviatedString());
-            return;
-          }
-
-          final int validatorIndex = maybeValidatorIndex.get();
-          final Validator validator = state.getValidators().get(validatorIndex);
-
-          // Check if validator has an execution address set
-          final boolean hasExecutionAddress =
-              predicatesElectra.hasExecutionWithdrawalCredential(validator);
-          if (!hasExecutionAddress) {
-            LOG.debug(
-                "process_withdrawal_request: validator index {} does not have withdrawal credentials set",
-                validatorIndex);
-            return;
-          }
-
-          // Check withdrawalRequest source_address matches validator eth1 withdrawal credentials
-          final Bytes20 validatorExecutionAddress =
-              new Bytes20(validator.getWithdrawalCredentials().slice(12));
-          final Bytes20 withdrawalRequestSourceAddress = withdrawalRequest.getSourceAddress();
-          final boolean isCorrectSourceAddress =
-              validatorExecutionAddress.equals(withdrawalRequestSourceAddress);
-          if (!isCorrectSourceAddress) {
-            LOG.debug(
-                "process_withdrawal_request: WithdrawalRequest source_address {} does not match "
-                    + "validator {} withdrawal credentials {}",
-                withdrawalRequestSourceAddress,
-                validatorIndex,
-                validatorExecutionAddress);
-            return;
-          }
-
-          // Check if validator is active
-          final boolean isValidatorActive = predicates.isActiveValidator(validator, currentEpoch);
-          if (!isValidatorActive) {
-            LOG.debug("process_withdrawal_request: Validator {} is not active", validatorIndex);
-            return;
-          }
-
-          // Check if validator has already initiated exit
-          final boolean hasInitiatedExit = !validator.getExitEpoch().equals(FAR_FUTURE_EPOCH);
-          if (hasInitiatedExit) {
-            LOG.debug(
-                "process_withdrawal_request: Validator {} has already initiated exit",
-                validatorIndex);
-            return;
-          }
-
-          // Check if validator has been active long enough
-          final boolean validatorActiveLongEnough =
-              currentEpoch.isGreaterThanOrEqualTo(
-                  validator.getActivationEpoch().plus(specConfig.getShardCommitteePeriod()));
-          if (!validatorActiveLongEnough) {
-            LOG.debug(
-                "process_withdrawal_request: Validator {} is not active long enough",
-                validatorIndex);
-            return;
-          }
-
-          final UInt64 pendingBalanceToWithdraw =
-              validatorsUtil.getPendingBalanceToWithdraw(state, validatorIndex);
-          if (isFullExitRequest) {
-            // Only exit validator if it has no pending withdrawals in the queue
-            if (pendingBalanceToWithdraw.isZero()) {
-              LOG.debug(
-                  "process_withdrawal_request: Initiating exit for validator {}", validatorIndex);
-
-              beaconStateMutators.initiateValidatorExit(
-                  state, validatorIndex, validatorExitContextSupplier);
-            }
-            return;
-          }
-
-          final UInt64 validatorBalance = state.getBalances().get(validatorIndex).get();
-          final UInt64 minActivationBalance = specConfigElectra.getMinActivationBalance();
-
-          final boolean hasCompoundingWithdrawalCredential =
-              predicatesElectra.hasCompoundingWithdrawalCredential(validator);
-          final boolean hasSufficientEffectiveBalance =
-              validator.getEffectiveBalance().isGreaterThanOrEqualTo(minActivationBalance);
-          final boolean hasExcessBalance =
-              validatorBalance.isGreaterThan(minActivationBalance.plus(pendingBalanceToWithdraw));
-          if (hasCompoundingWithdrawalCredential
-              && hasSufficientEffectiveBalance
-              && hasExcessBalance) {
-            final UInt64 toWithdraw =
-                validatorBalance
-                    .minusMinZero(minActivationBalance)
-                    .minusMinZero(pendingBalanceToWithdraw)
-                    .min(withdrawalRequest.getAmount());
-            final MutableBeaconStateElectra electraState =
-                MutableBeaconStateElectra.required(state);
-            final UInt64 exitQueueEpoch =
-                beaconStateMutatorsElectra.computeExitEpochAndUpdateChurn(electraState, toWithdraw);
-            final UInt64 withdrawableEpoch =
-                exitQueueEpoch.plus(specConfigElectra.getMinValidatorWithdrawabilityDelay());
-
-            LOG.debug(
-                "process_withdrawal_request: Creating pending partial withdrawal for validator {}",
-                validatorIndex);
-
-            electraState
-                .getPendingPartialWithdrawals()
-                .append(
-                    schemaDefinitionsElectra
-                        .getPendingPartialWithdrawalSchema()
-                        .create(
-                            SszUInt64.of(UInt64.fromLongBits(validatorIndex)),
-                            SszUInt64.of(toWithdraw),
-                            SszUInt64.of(withdrawableEpoch)));
-          }
-        });
-  }
-
-  /**
-   * Implements process_consolidation_request from consensus-spec (EIP-7251)
-   *
-   * @see <a
-   *     href="https://github.com/ethereum/consensus-specs/blob/master/specs/electra/beacon-chain
-   *     .md#new-process_consolidation_request"/>
-   */
-  @Override
-  public void processConsolidationRequests(
-      final MutableBeaconState state, final List<ConsolidationRequest> consolidationRequests) {
-    LOG.debug(
-        "process_consolidation_request: {} consolidation requests to process from block at "
-            + "slot {}",
-        consolidationRequests.size(),
-        state.getSlot());
-
-    final MutableBeaconStateElectra electraState = MutableBeaconStateElectra.required(state);
-    consolidationRequests.forEach(
-        consolidationRequest -> processConsolidationRequest(electraState, consolidationRequest));
-  }
-
-  private void processConsolidationRequest(
-      final MutableBeaconStateElectra state, final ConsolidationRequest consolidationRequest) {
-    final UInt64 slot = state.getSlot();
-    final UInt64 currentEpoch = miscHelpers.computeEpochAtSlot(slot);
-
-    if (isValidSwitchToCompoundingRequest(state, consolidationRequest)) {
-      LOG.debug(
-          "process_consolidation_request: switching validator {} to compounding address",
-          consolidationRequest.getSourcePubkey().toAbbreviatedString());
-      validatorsUtil
-          .getValidatorIndex(state, consolidationRequest.getSourcePubkey())
-          .ifPresent(
-              sourceValidatorIndex ->
-                  beaconStateMutatorsElectra.switchToCompoundingValidator(
-                      state, sourceValidatorIndex));
-      return;
-    }
-
-    // Verify that source != target, so a consolidation cannot be used as an exit
-    if (consolidationRequest.getSourcePubkey().equals(consolidationRequest.getTargetPubkey())) {
-      LOG.debug(
-          "process_consolidation_request: source_pubkey and target_pubkey must be different (pubkey = {})",
-          consolidationRequest.getSourcePubkey().toAbbreviatedString());
-      return;
-    }
-
-    // If the pending consolidations queue is full, consolidation requests are ignored
-    if (state.getPendingConsolidations().size()
-        == specConfigElectra.getPendingConsolidationsLimit()) {
-      LOG.debug("process_consolidation_request: consolidation queue is full");
-      return;
-    }
-
-    // If there is too little available consolidation churn limit, consolidation requests are
-    // ignored
-    if (beaconStateAccessorsElectra
-        .getConsolidationChurnLimit(state)
-        .isLessThanOrEqualTo(specConfigElectra.getMinActivationBalance())) {
-      LOG.debug("process_consolidation_request: not enough consolidation churn limit available");
-      return;
-    }
-
-    // Verify source_pubkey exists
-    final Optional<Integer> maybeSourceValidatorIndex =
-        validatorsUtil.getValidatorIndex(state, consolidationRequest.getSourcePubkey());
-    if (maybeSourceValidatorIndex.isEmpty()) {
-      LOG.debug(
-          "process_consolidation_request: source_pubkey {} not found",
-          consolidationRequest.getSourcePubkey().toAbbreviatedString());
-      return;
-    }
-
-    // Verify target_pubkey exists
-    final Optional<Integer> maybeTargetValidatorIndex =
-        validatorsUtil.getValidatorIndex(state, consolidationRequest.getTargetPubkey());
-    if (maybeTargetValidatorIndex.isEmpty()) {
-      LOG.debug(
-          "process_consolidation_request: target_pubkey {} not found",
-          consolidationRequest.getTargetPubkey().toAbbreviatedString());
-      return;
-    }
-
-    final int sourceValidatorIndex = maybeSourceValidatorIndex.get();
-    final Validator sourceValidator = state.getValidators().get(sourceValidatorIndex);
-    final int targetValidatorIndex = maybeTargetValidatorIndex.get();
-    final Validator targetValidator = state.getValidators().get(targetValidatorIndex);
-
-    // Verify source withdrawal credentials
-    final boolean sourceHasExecutionWithdrawalCredentials =
-        predicatesElectra.hasExecutionWithdrawalCredential(sourceValidator);
-
-    final Eth1Address sourceValidatorExecutionAddress =
-        Predicates.getExecutionAddressUnchecked(sourceValidator.getWithdrawalCredentials());
-    final boolean sourceHasCorrectCredentials =
-        sourceValidatorExecutionAddress.equals(
-            Eth1Address.fromBytes(consolidationRequest.getSourceAddress().getWrappedBytes()));
-    if (!(sourceHasExecutionWithdrawalCredentials && sourceHasCorrectCredentials)) {
-      LOG.debug("process_consolidation_request: invalid source credentials");
-      return;
-    }
-
-    // Verify that target has compounding withdrawal credentials
-    if (!predicatesElectra.hasCompoundingWithdrawalCredential(targetValidator)) {
-      LOG.debug("process_consolidation_request: invalid target credentials");
-      return;
-    }
-
-    // Verify the source and the target are active
-    if (!predicatesElectra.isActiveValidator(sourceValidator, currentEpoch)) {
-      LOG.debug(
-          "process_consolidation_request: source validator {} is inactive", sourceValidatorIndex);
-      return;
-    }
-    if (!predicatesElectra.isActiveValidator(targetValidator, currentEpoch)) {
-      LOG.debug(
-          "process_consolidation_request: target validator {} is inactive", targetValidatorIndex);
-      return;
-    }
-
-    // Verify exits for source and target have not been initiated
-    if (!sourceValidator.getExitEpoch().equals(FAR_FUTURE_EPOCH)) {
-      LOG.debug(
-          "process_consolidation_request: source validator {} is exiting", sourceValidatorIndex);
-      return;
-    }
-    if (!targetValidator.getExitEpoch().equals(FAR_FUTURE_EPOCH)) {
-      LOG.debug(
-          "process_consolidation_request: target validator {} is exiting", targetValidatorIndex);
-      return;
-    }
-
-    // Verify the source has been active long enough
-    if (currentEpoch.isLessThan(
-        sourceValidator.getActivationEpoch().plus(specConfig.getShardCommitteePeriod()))) {
-      LOG.debug("process_consolidation_request: source has not been active long enough");
-      return;
-    }
-    // Verify the source has no pending withdrawals in the queue
-    if (beaconStateAccessorsElectra
-        .getPendingBalanceToWithdraw(state, sourceValidatorIndex)
-        .isGreaterThan(ZERO)) {
-      LOG.debug("process_consolidation_request: source has pending withdrawals in the queue");
-      return;
-    }
-
-    // Initiate source validator exit and append pending consolidation
-    final UInt64 exitEpoch =
-        beaconStateMutatorsElectra.computeConsolidationEpochAndUpdateChurn(
-            state, sourceValidator.getEffectiveBalance());
-    final UInt64 withdrawableEpoch =
-        exitEpoch.plus(specConfigElectra.getMinValidatorWithdrawabilityDelay());
-
-    state
-        .getValidators()
-        .update(
-            sourceValidatorIndex,
-            v -> v.withExitEpoch(exitEpoch).withWithdrawableEpoch(withdrawableEpoch));
-    LOG.debug(
-        "process_consolidation_request: updated validator {} with exit_epoch = {}, withdrawable_epoch = {}",
-        sourceValidatorIndex,
-        exitEpoch,
-        withdrawableEpoch);
-
-    final PendingConsolidation pendingConsolidation =
-        new PendingConsolidation(
-            schemaDefinitionsElectra.getPendingConsolidationSchema(),
-            SszUInt64.of(UInt64.valueOf(sourceValidatorIndex)),
-            SszUInt64.of(UInt64.valueOf(targetValidatorIndex)));
-    state.getPendingConsolidations().append(pendingConsolidation);
-
-    LOG.debug("process_consolidation_request: created {}", pendingConsolidation);
-  }
-
-  /**
-   * Implements function is_valid_switch_to_compounding_request
-   *
-   * @see <a
-   *     href="https://github.com/ethereum/consensus-specs/blob/master/specs/electra/beacon-chain.md#new-is_valid_switch_to_compounding_request"/>
-   */
-  @Override
-  public boolean isValidSwitchToCompoundingRequest(
-      final BeaconState state, final ConsolidationRequest consolidationRequest) {
-
-    // Switch to compounding requires source and target be equal
-    if (!consolidationRequest.getSourcePubkey().equals(consolidationRequest.getTargetPubkey())) {
-      return false;
-    }
-
-    // Verify source_pubkey exists
-    final Optional<Integer> maybeSourceValidatorIndex =
-        validatorsUtil.getValidatorIndex(state, consolidationRequest.getSourcePubkey());
-    if (maybeSourceValidatorIndex.isEmpty()) {
-      return false;
-    }
-
-    final int sourceValidatorIndex = maybeSourceValidatorIndex.get();
-    final Validator sourceValidator = state.getValidators().get(sourceValidatorIndex);
-
-    // Verify request has been authorized
-    final Eth1Address sourceValidatorExecutionAddress =
-        Predicates.getExecutionAddressUnchecked(sourceValidator.getWithdrawalCredentials());
-    if (!sourceValidatorExecutionAddress.equals(
-        Eth1Address.fromBytes(consolidationRequest.getSourceAddress().getWrappedBytes()))) {
-      return false;
-    }
-
-    // Verify source withdrawal credentials
-    if (!predicatesElectra.hasEth1WithdrawalCredential(sourceValidator)) {
-      return false;
-    }
-
-    // Verify the source is active
-    final UInt64 currentEpoch = miscHelpers.computeEpochAtSlot(state.getSlot());
-    if (!predicatesElectra.isActiveValidator(sourceValidator, currentEpoch)) {
-      return false;
-    }
-
-    // Verify exit for source has not been initiated
-    return sourceValidator.getExitEpoch().equals(FAR_FUTURE_EPOCH);
   }
 
   @Override
