+++ ethereum/spec/src/main/java/tech/pegasys/teku/spec/logic/versions/fulu/block/BlockProcessorFulu.java
@@ -24,6 +24,7 @@ import tech.pegasys.teku.spec.logic.common.util.BeaconStateUtil;
 import tech.pegasys.teku.spec.logic.common.util.SyncCommitteeUtil;
 import tech.pegasys.teku.spec.logic.common.util.ValidatorsUtil;
 import tech.pegasys.teku.spec.logic.versions.electra.block.BlockProcessorElectra;
+import tech.pegasys.teku.spec.logic.versions.electra.execution.ExecutionRequestsProcessorElectra;
 import tech.pegasys.teku.spec.logic.versions.electra.helpers.BeaconStateAccessorsElectra;
 import tech.pegasys.teku.spec.logic.versions.electra.helpers.BeaconStateMutatorsElectra;
 import tech.pegasys.teku.spec.logic.versions.fulu.helpers.MiscHelpersFulu;
@@ -45,7 +46,8 @@ public class BlockProcessorFulu extends BlockProcessorElectra {
       final ValidatorsUtil validatorsUtil,
       final OperationValidator operationValidator,
       final SchemaDefinitionsElectra schemaDefinitions,
-      final ExecutionRequestsDataCodec executionRequestsDataCodec) {
+      final ExecutionRequestsDataCodec executionRequestsDataCodec,
+      final ExecutionRequestsProcessorElectra executionRequestsProcessor) {
     super(
         specConfig,
         predicates,
@@ -59,7 +61,8 @@ public class BlockProcessorFulu extends BlockProcessorElectra {
         validatorsUtil,
         operationValidator,
         schemaDefinitions,
-        executionRequestsDataCodec);
+        executionRequestsDataCodec,
+        executionRequestsProcessor);
     miscHelpersFulu = miscHelpers;
   }
