+++ ethereum/spec/src/test/java/tech/pegasys/teku/spec/logic/common/block/BlockProcessorTest.java
@@ -17,8 +17,6 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
-import java.util.ArrayList;
-import java.util.List;
 import org.apache.tuweni.bytes.Bytes;
 import org.apache.tuweni.bytes.Bytes32;
 import org.apache.tuweni.bytes.Bytes48;
@@ -30,32 +28,21 @@ import tech.pegasys.teku.infrastructure.ssz.SszList;
 import tech.pegasys.teku.infrastructure.ssz.collections.SszBytes32Vector;
 import tech.pegasys.teku.infrastructure.ssz.schema.SszListSchema;
 import tech.pegasys.teku.infrastructure.unsigned.UInt64;
-import tech.pegasys.teku.spec.Spec;
 import tech.pegasys.teku.spec.SpecMilestone;
-import tech.pegasys.teku.spec.SpecVersion;
-import tech.pegasys.teku.spec.config.SpecConfig;
 import tech.pegasys.teku.spec.datastructures.blocks.BeaconBlockHeader;
 import tech.pegasys.teku.spec.datastructures.blocks.Eth1Data;
 import tech.pegasys.teku.spec.datastructures.operations.Deposit;
 import tech.pegasys.teku.spec.datastructures.operations.DepositData;
 import tech.pegasys.teku.spec.datastructures.operations.DepositMessage;
-import tech.pegasys.teku.spec.datastructures.state.Fork;
 import tech.pegasys.teku.spec.datastructures.state.Validator;
 import tech.pegasys.teku.spec.datastructures.state.beaconstate.BeaconState;
 import tech.pegasys.teku.spec.datastructures.state.beaconstate.MutableBeaconState;
 import tech.pegasys.teku.spec.datastructures.util.MerkleTree;
+import tech.pegasys.teku.spec.logic.common.ProcessorTestHelper;
 import tech.pegasys.teku.spec.logic.common.statetransition.exceptions.BlockProcessingException;
-import tech.pegasys.teku.spec.util.DataStructureUtil;
 
-public abstract class BlockProcessorTest {
-  protected final Spec spec = createSpec();
-  protected final DataStructureUtil dataStructureUtil = new DataStructureUtil(spec);
-
-  private final SpecVersion genesisSpec = spec.getGenesisSpec();
-  protected final SpecConfig specConfig = genesisSpec.getConfig();
+public abstract class BlockProcessorTest extends ProcessorTestHelper {
   private final BlockProcessor blockProcessor = genesisSpec.getBlockProcessor();
-
-  protected abstract Spec createSpec();
 
   @Test
   void ensureDepositSignatureVerifierHasDefaultValue() {
@@ -225,52 +212,6 @@ public abstract class BlockProcessorTest {
         "The balances list has changed.");
   }
 
-  protected BeaconState createBeaconState() {
-    return createBeaconState(false, null, null);
-  }
-
-  protected BeaconState createBeaconState(final UInt64 amount, final Validator knownValidator) {
-    return createBeaconState(true, amount, knownValidator);
-  }
-
-  protected BeaconState createBeaconState(
-      final boolean addToList, final UInt64 amount, final Validator knownValidator) {
-    return spec.getGenesisSpec()
-        .getSchemaDefinitions()
-        .getBeaconStateSchema()
-        .createEmpty()
-        .updated(
-            beaconState -> {
-              beaconState.setSlot(dataStructureUtil.randomUInt64());
-              beaconState.setFork(
-                  new Fork(
-                      specConfig.getGenesisForkVersion(),
-                      specConfig.getGenesisForkVersion(),
-                      SpecConfig.GENESIS_EPOCH));
-
-              List<Validator> validatorList =
-                  new ArrayList<>(
-                      List.of(
-                          dataStructureUtil.randomValidator(),
-                          dataStructureUtil.randomValidator(),
-                          dataStructureUtil.randomValidator()));
-              List<UInt64> balanceList =
-                  new ArrayList<>(
-                      List.of(
-                          dataStructureUtil.randomUInt64(),
-                          dataStructureUtil.randomUInt64(),
-                          dataStructureUtil.randomUInt64()));
-
-              if (addToList) {
-                validatorList.add(knownValidator);
-                balanceList.add(amount);
-              }
-
-              beaconState.getValidators().appendAll(validatorList);
-              beaconState.getBalances().appendAllElements(balanceList);
-            });
-  }
-
   protected BeaconState processDepositHelper(
       final BeaconState beaconState, final DepositData depositData)
       throws BlockProcessingException {
@@ -292,27 +233,5 @@ public abstract class BlockProcessorTest {
 
     // Attempt to process deposit with above data.
     return updatedBeaconState.updated(state -> blockProcessor.processDeposits(state, deposits));
-  }
-
-  protected Validator makeValidator(
-      final BLSPublicKey pubkey, final Bytes32 withdrawalCredentials) {
-    return makeValidator(
-        pubkey, withdrawalCredentials, SpecConfig.FAR_FUTURE_EPOCH, SpecConfig.FAR_FUTURE_EPOCH);
-  }
-
-  protected Validator makeValidator(
-      final BLSPublicKey pubkey,
-      final Bytes32 withdrawalCredentials,
-      final UInt64 exitEpoch,
-      final UInt64 withdrawableEpoch) {
-    return new Validator(
-        pubkey,
-        withdrawalCredentials,
-        specConfig.getMaxEffectiveBalance(),
-        false,
-        SpecConfig.FAR_FUTURE_EPOCH,
-        SpecConfig.FAR_FUTURE_EPOCH,
-        exitEpoch,
-        withdrawableEpoch);
   }
 }
